Configure button to control axis and color variables.
The Highcharts plot gets unhappy when it's inside of a v-app, so the v-app only contains the configure button. This is a bit hacky, but it works for now. The button has the interface that we want, but changing things there doesn't change the plot properly; we'll merge in Rachel's changes to just redo the plot.